requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==1.12.0
python-dotenv==0.19.0
werkzeug==2.0.1
httpx==0.24.1 